requirements/local.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.5
django-extensions==3.1.5
djangorestframework==3.13.1
gunicorn==20.1.0
kombu==5.2.4
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
psycopg2